+++ Cargo.toml
@@ -10,7 +10,6 @@ members = [
   "crates/next-build",
   "crates/next-core",
   "crates/next-custom-transforms",
-  "crates/next-persistence-tools",
   "crates/next-taskless",
   "turbopack/crates/*",
   "turbopack/crates/*/fuzz",
